# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/library": {
      "get": {
        "operationId": "LibraryController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "E-books acquis, avec la progression de lecture",
        "tags": [
          "Bibliothèque"
        ]
      }
    },
    "/v1/library/{ebookId}/content": {
      "get": {
        "operationId": "LibraryController_content",
        "parameters": [
          {
            "name": "ebookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Contenu complet pour le lecteur (chapitres en blocs de texte)",
        "tags": [
          "Bibliothèque"
        ]
      }
    },
    "/v1/library/{ebookId}/download-url": {
      "post": {
        "operationId": "LibraryController_downloadUrl",
        "parameters": [
          {
            "name": "ebookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "URL signée et temporaire du fichier original, si le téléchargement est autorisé",
        "tags": [
          "Bibliothèque"
        ]
      }
    },
    "/v1/library/{ebookId}/progress": {
      "put": {
        "operationId": "LibraryController_progress",
        "parameters": [
          {
            "name": "ebookId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enregistrer la progression (dernier écrit gagne)",
        "tags": [
          "Bibliothèque"
        ]
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "MeController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Compte"
        ]
      },
      "patch": {
        "operationId": "MeController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Compte"
        ]
      },
      "delete": {
        "operationId": "MeController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprimer le compte : anonymisation, commandes conservées (obligation comptable)",
        "tags": [
          "Compte"
        ]
      }
    },
    "/v1/me/password": {
      "put": {
        "operationId": "MeController_password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Changer de mot de passe (ferme les autres sessions)",
        "tags": [
          "Compte"
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "operationId": "MeController_notifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Compte"
        ]
      },
      "put": {
        "operationId": "MeController_setNotifications",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Compte"
        ]
      }
    },
    "/v1/me/devices": {
      "get": {
        "operationId": "MeController_devices",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Compte"
        ]
      }
    },
    "/v1/me/devices/{id}": {
      "delete": {
        "operationId": "MeController_removeDevice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retirer un appareil : ses sessions sont fermées",
        "tags": [
          "Compte"
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "operationId": "MeController_export",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export RGPD de toutes les données du compte",
        "tags": [
          "Compte"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Folio",
    "description": "Vente et lecture d'e-books — application mobile et administration. Toutes les routes sont préfixées par /v1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProgressDto": {
        "type": "object",
        "properties": {
          "chapterIndex": {
            "type": "number"
          },
          "blockIndex": {
            "type": "number"
          },
          "percent": {
            "type": "number"
          }
        },
        "required": [
          "chapterIndex",
          "blockIndex",
          "percent"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "NotificationsDto": {
        "type": "object",
        "properties": {
          "ebookUpdates": {
            "type": "boolean",
            "description": "Mises à jour d'e-books déjà achetés"
          },
          "marketing": {
            "type": "boolean",
            "description": "Nouveautés et offres (consentement)"
          }
        },
        "required": [
          "ebookUpdates",
          "marketing"
        ]
      },
      "DeleteAccountDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password"
        ]
      },
...
    }
  }
}